+++ packages/core/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Fix module not found error. ([#235](https://github.com/OpenZeppelin/contracts-wizard/issues/235))
+
 ## 0.2.2 (2023-03-17)
 
 - Fix missing file. ([#234](https://github.com/OpenZeppelin/contracts-wizard/pull/234))